Move parameter files
