+++ src/coreComponents/constitutive/fluid/multifluid/compositional/functions/StabilityTest.hpp
@@ -24,7 +24,7 @@
 #include "FugacityCalculator.hpp"
 #include "constitutive/fluid/multifluid/Layouts.hpp"
 #include "constitutive/fluid/multifluid/MultiFluidConstants.hpp"
-#include "constitutive/fluid/multifluid/compositional/models/ComponentProperties.hpp"
+#include "constitutive/fluid/multifluid/compositional/parameters/ComponentProperties.hpp"
 
 namespace geos
 {
