+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/ImmiscibleWaterDensity.cpp
@@ -18,7 +18,7 @@
  */
 
 #include "ImmiscibleWaterDensity.hpp"
-#include "ImmiscibleWaterParameters.hpp"
+#include "constitutive/fluid/multifluid/compositional/parameters/ImmiscibleWaterParameters.hpp"
 #include "dataRepository/InputFlags.hpp"
 
 namespace geos
