+++ src/coreComponents/constitutive/fluid/multifluid/compositional/parameters/ComponentProperties.hpp
@@ -17,8 +17,8 @@
  * @file ComponentProperties.hpp
  */
 
-#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_COMPONENTPROPERTIES_HPP_
-#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_COMPONENTPROPERTIES_HPP_
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_COMPONENTPROPERTIES_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_COMPONENTPROPERTIES_HPP_
 
 #include "common/DataTypes.hpp"
 
@@ -136,4 +136,4 @@ public:
 
 } // namespace geos
 
-#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_COMPONENTPROPERTIES_HPP_
+#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_COMPONENTPROPERTIES_HPP_
