+++ src/coreComponents/constitutive/fluid/multifluid/compositional/parameters/CriticalVolume.hpp
@@ -17,8 +17,8 @@
  * @file CriticalVolume.hpp
  */
 
-#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_CRITICALVOLUME_HPP_
-#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_CRITICALVOLUME_HPP_
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_CRITICALVOLUME_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_CRITICALVOLUME_HPP_
 
 #include "ModelParameters.hpp"
 #include "common/DataTypes.hpp"
@@ -72,4 +72,4 @@ protected:
 
 } // end namespace geos
 
-#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_CRITICALVOLUME_HPP_
+#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_CRITICALVOLUME_HPP_
