+++ src/coreComponents/constitutive/fluid/multifluid/compositional/parameters/EquationOfState.hpp
@@ -17,8 +17,8 @@
  * @file EquationOfState.hpp
  */
 
-#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_EQUATIONOFSTATE_HPP_
-#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_EQUATIONOFSTATE_HPP_
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_EQUATIONOFSTATE_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_EQUATIONOFSTATE_HPP_
 
 #include "ModelParameters.hpp"
 #include "constitutive/fluid/multifluid/MultiFluidBase.hpp"
@@ -103,4 +103,4 @@ protected:
 
 } // end namespace geos
 
-#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_EQUATIONOFSTATE_HPP_
+#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_EQUATIONOFSTATE_HPP_
