+++ src/coreComponents/constitutive/fluid/multifluid/compositional/parameters/ImmiscibleWaterParameters.hpp
@@ -17,8 +17,8 @@
  * @file ImmiscibleWaterParameters.hpp
  */
 
-#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_IMMISCIBLEWATERPARAMETERS_HPP_
-#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_IMMISCIBLEWATERPARAMETERS_HPP_
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_IMMISCIBLEWATERPARAMETERS_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_IMMISCIBLEWATERPARAMETERS_HPP_
 
 #include "ModelParameters.hpp"
 #include "common/DataTypes.hpp"
@@ -77,4 +77,4 @@ protected:
 
 } // end namespace geos
 
-#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_IMMISCIBLEWATERPARAMETERS_HPP_
+#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_IMMISCIBLEWATERPARAMETERS_HPP_
