+++ src/coreComponents/constitutive/fluid/multifluid/compositional/parameters/ModelParameters.hpp
@@ -17,8 +17,8 @@
  * @file ModelParameters.hpp
  */
 
-#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_MODELPARAMETERS_HPP_
-#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_MODELPARAMETERS_HPP_
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_MODELPARAMETERS_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_MODELPARAMETERS_HPP_
 
 #include <memory>
 #include "common/GeosxMacros.hpp"
@@ -93,4 +93,4 @@ private:
 
 } // end namespace geos
 
-#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_MODELS_MODELPARAMETERS_HPP_
+#endif //GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_PARAMETERS_MODELPARAMETERS_HPP_
